Remove routes component in console
