+++ frontend/packages/topology/src/components/workload/NetworkingOverview.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { List, ListItem } from '@patternfly/react-core';
 import { LongArrowAltRightIcon } from '@patternfly/react-icons/dist/esm/icons/long-arrow-alt-right-icon';
 import { useTranslation } from 'react-i18next';
-import { RouteLocation } from '@console/internal/components/routes';
 import { ResourceLink, SidebarSectionHeading } from '@console/internal/components/utils';
 import { K8sResourceKind, RouteKind } from '@console/internal/module/k8s';
 import { useRoutesWatcher, useServicesWatcher } from '@console/shared';
+import { RouteLocation } from '@console/shared/src/components/utils/routes';
 
 const ServicePortList: React.FC<ServicePortListProps> = ({ service }) => {
   const ports = service.spec?.ports ?? [];
