+++ frontend/public/locales/en/public.json
@@ -1275,70 +1275,6 @@
   "Cluster-wide {{resource}}": "Cluster-wide {{resource}}",
   "Namespace {{resource}}": "Namespace {{resource}}",
   "Create ResourceQuota": "Create ResourceQuota",
-  "unknown host": "unknown host",
-  "TLS is not enabled": "TLS is not enabled",
-  "Termination type": "Termination type",
-  "Insecure traffic": "Insecure traffic",
-  "Hide": "Hide",
-  "Reveal": "Reveal",
-  "CA certificate": "CA certificate",
-  "Destination CA certificate": "Destination CA certificate",
-  "Custom route": "Custom route",
-  "To use a custom route, you must update your DNS provider by creating a canonical name (CNAME) record. Your CNAME record should point to your custom domain <2>{{host}}</2>, to the OpenShift canonical router hostname, <5>{{routerCanonicalHostname}}</5>, as the alias.": "To use a custom route, you must update your DNS provider by creating a canonical name (CNAME) record. Your CNAME record should point to your custom domain <2>{{host}}</2>, to the OpenShift canonical router hostname, <5>{{routerCanonicalHostname}}</5>, as the alias.",
-  "Do you need to set up custom DNS?": "Do you need to set up custom DNS?",
-  "Router: {{routerName}}": "Router: {{routerName}}",
-  "Host": "Host",
-  "Wildcard policy": "Wildcard policy",
-  "Router canonical hostname": "Router canonical hostname",
-  "Route details": "Route details",
-  "Service": "Service",
-  "Target port": "Target port",
-  "Path": "Path",
-  "TLS settings": "TLS settings",
-  "Traffic": "Traffic",
-  "This route splits traffic across multiple services.": "This route splits traffic across multiple services.",
-  "Weight": "Weight",
-  "Percent": "Percent",
-  "No route status": "No route status",
-  "Routes": "Routes",
-  "Accepted": "Accepted",
-  "Rejected": "Rejected",
-  "Remove alternate Service": "Remove alternate Service",
-  "A unique name for the Route within the project.": "A unique name for the Route within the project.",
-  "Public hostname for the Route. If not specified, a hostname is generated.": "Public hostname for the Route. If not specified, a hostname is generated.",
-  "Path that the router watches to route traffic to the service.": "Path that the router watches to route traffic to the service.",
-  "There are no Services in your project to expose with a Route.": "There are no Services in your project to expose with a Route.",
-  "Select a service": "Select a service",
-  "Service to route to.": "Service to route to.",
-  "A number between 0 and 255 that depicts relative weight compared with other targets.": "A number between 0 and 255 that depicts relative weight compared with other targets.",
-  "Add alternate Service": "Add alternate Service",
-  "Select a Service above": "Select a Service above",
-  "Select target port": "Select target port",
-  "Target port for traffic.": "Target port for traffic.",
-  "Security": "Security",
-  "Secure Route": "Secure Route",
-  "Routes can be secured using several TLS termination types for serving certificates.": "Routes can be secured using several TLS termination types for serving certificates.",
-  "TLS termination": "TLS termination",
-  "Select termination type": "Select termination type",
-  "Select insecure traffic type": "Select insecure traffic type",
-  "Policy for traffic on insecure schemes like HTTP.": "Policy for traffic on insecure schemes like HTTP.",
-  "Certificates": "Certificates",
-  "TLS certificates for edge and re-encrypt termination. If not specified, the router&apos;s default certificate is used.": "TLS certificates for edge and re-encrypt termination. If not specified, the router&apos;s default certificate is used.",
-  "The PEM format certificate. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.": "The PEM format certificate. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.",
-  "Private key": "Private key",
-  "The PEM format key. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.": "The PEM format key. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.",
-  "The PEM format CA certificate chain. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.": "The PEM format CA certificate chain. Upload file by dragging & dropping, selecting it, or pasting from the clipboard.",
-  "Alternate Service target": "Alternate Service target",
-  "Alternate Service to route to.": "Alternate Service to route to.",
-  "Alternate Service weight": "Alternate Service weight",
-  "Traffic in": "Traffic in",
-  "Traffic out": "Traffic out",
-  "Connection rate": "Connection rate",
-  "Routing is a way to make your application publicly visible.": "Routing is a way to make your application publicly visible.",
-  "Edit Route": "Edit Route",
-  "Create Route": "Create Route",
-  "Insufficient permissions to update host.": "Insufficient permissions to update host.",
-  "Unknown error submitting": "Unknown error submitting",
   "Select all filters": "Select all filters",
   "{{selectedCount}} of {{itemCount}}": "{{selectedCount}} of {{itemCount}}",
   "Item_one": "Item",
@@ -1632,6 +1568,7 @@
   "ServiceMonitors": "ServiceMonitors",
   "PodMonitor": "PodMonitor",
   "PodMonitors": "PodMonitors",
+  "Service": "Service",
   "Services": "Services",
   "DaemonSet": "DaemonSet",
   "ReplicationController": "ReplicationController",
@@ -1659,6 +1596,7 @@
   "Ingress": "Ingress",
   "Ingresses": "Ingresses",
   "Route": "Route",
+  "Routes": "Routes",
   "ConfigMap": "ConfigMap",
   "ClusterRoleBinding": "ClusterRoleBinding",
   "ClusterRoleBindings": "ClusterRoleBindings",
@@ -1801,6 +1739,7 @@
   "Query result is a string, which cannot be graphed.": "Query result is a string, which cannot be graphed.",
   "The resulting dataset is too large to graph.": "The resulting dataset is too large to graph.",
   "Stacked": "Stacked",
+  "unknown host": "unknown host",
   "Just now": "Just now",
   "Disable": "Disable",
   "Disabled": "Disabled",
