+++ frontend/public/style.scss
@@ -80,7 +80,6 @@
 @import 'components/overview/overview';
 @import 'components/deploy-image';
 @import 'components/instantiate-template';
-@import 'components/routes/create-route';
 @import 'components/storage-class-form';
 @import 'components/quota';
 @import 'components/cluster-settings/cluster-settings';
